feat[#231] : 이전 image 서버에 있던 부분 하나로 통합
- 이미지 서버를 따로 두지않고 서버에 정적파일들을 저장해서 사용하기로 팀원들과의 협의끝에 결론이 나서, 다시 server의 api/chat/upload/:postId로 이미지를 업로드할 수 있도록 만들었다.
- 이미지는 upload 폴더의 각각 채팅방 이름에 따라 만들어진다.
