--- imageServer/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "imageserver",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "npm run build:live",
-    "build:live": "nodemon --watch '*.ts' --exec ts-node app.ts",
-    "build": "tsc -p ./"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@types/cors": "^2.8.12",
-    "@types/multer": "^1.4.7",
-    "cors": "^2.8.5",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "multer": "^1.4.3",
-    "sharp": "^0.29.3",
-    "ts-node": "^10.4.0"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/node": "^16.11.9",
-    "@types/sharp": "^0.29.4",
-    "nodemon": "^2.0.15",
-    "typescript": "^4.5.2"
-  }
-}
