--- imageServer/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"compilerOptions": {
-		"experimentalDecorators": true,
-		"target": "es6" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
-		"module": "commonjs" /* Specify what module code is generated. */,
-		"esModuleInterop": true /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables `allowSyntheticDefaultImports` for type compatibility. */,
-		"sourceMap": true,
-		"outDir": "./dist",
-		"rootDir": "./",
-		"strict": true /* Enable all strict type-checking options. */,
-		"moduleResolution": "node",
-		"resolveJsonModule": true
-	},
-	"exclude": ["node_modules"]
-}
