+++ server/package.json
@@ -13,14 +13,17 @@
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
+		"@types/multer": "^1.4.7",
 		"cookie-parser": "^1.4.5",
 		"cors": "^2.8.5",
 		"dotenv": "^10.0.0",
 		"express": "^4.17.1",
 		"html-metadata-parser": "^2.0.2",
 		"jsonwebtoken": "^8.5.1",
+		"multer": "^1.4.3",
 		"mysql2": "^2.3.2",
 		"reflect-metadata": "^0.1.13",
+		"sharp": "^0.29.3",
 		"socket.io": "^4.3.2",
 		"typeorm": "^0.2.38"
 	},
@@ -30,6 +33,7 @@
 		"@types/express": "^4.17.13",
 		"@types/jsonwebtoken": "^8.5.5",
 		"@types/node": "^16.11.6",
+		"@types/sharp": "^0.29.4",
 		"nodemon": "^2.0.14",
 		"ts-node": "^10.4.0",
 		"typescript": "^4.4.4"
