+++ server/routes/chat.ts
@@ -3,12 +3,25 @@ import {
   getParticipants,
   createParticipant
 } from '../controller/participant-controller';
-import { getChats } from '../controller/chat-controller';
+import {
+  checkDirectory,
+  getChats,
+  resizeImg,
+  upload,
+  uploadImage
+} from '../controller/chat-controller';
 
 const router = express.Router();
 router.get('/:post_id', getChats);
 router.get('/:post_id/participant', getParticipants);
 router.post('/:post_id/participant', createParticipant);
+router.post(
+  '/upload/:postId',
+  checkDirectory,
+  upload.single('file'),
+  resizeImg,
+  uploadImage
+);
 
 const errorHandler = (err: any, req: any, res: any, next: any) => {
   res.status(err.statusCode).json(err.message);
